Current bug with pause option
When the game is paused and you click on the how to play button the combat mode and dance mode timer will keep going. This leads to combat mode overlapping with dance mode. I need a way to tell the game that it is still paused when the player is looking at the how to play menu in game.